norm 0.666

+++ src/obj_counters.c
@@ -3,8 +3,8 @@
 //1-A 2-C 3-L 4-sp 5-pl 6-cy 7-co
 int sum_one_type(int type, t_data *data, int *obj_types)
 {
-	int i;
-	int counter;
+	size_t i;
+	int 	counter;
 
 	counter = 0;
 	i = 0;
